Add basic structure
Added a class for the triager - the class that
will take care of pulling the data, processing it
and publishing it.

In addition added a class to handle the configuration
of the app. In the meantime, only reading it.

Updated the README. It doesn't reflect the current status
of the project, more of a roadmap as to where it should
be soon.